+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
 ## [0.1.7]
+- Adapt `in-any-order` to print diff for element ordering that leads to most direct matches
 - When matcher is provided with incorrect input, cause matcher to fail, but don't raise exception
 
 ## [0.1.6-SNAPSHOT]